Modify Item, item repository, item service Add item console
